Push Swap

+++ libftprintf/srcs/print_int.c
@@ -6,7 +6,7 @@
 /*   By: msantos2 <msantos2@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/30 16:00:48 by msantos2          #+#    #+#             */
-/*   Updated: 2025/12/03 15:08:31 by msantos2         ###   ########.fr       */
+/*   Updated: 2025/12/03 17:10:25 by msantos2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,13 @@ char	*int_conditions(t_flags *flag, char *nbr)
 	int	check;
 
 	check = 1;
-	if (flag->precision == -1 || flag->precision < ft_strlen(nbr))
+	if (flag->precision == -1 || flag->precision < (int)ft_strlen(nbr))
 		check = -1;
 	if (flag->precision != -1)
 		flag->zero = -1;
 	else
 		flag->precision = ft_strlen(nbr);
-	if (flag->precision < ft_strlen(nbr) && ft_atoi(nbr) != 0)
+	if (flag->precision < (int)ft_strlen(nbr) && ft_atoi(nbr) != 0)
 		flag->precision = ft_strlen(nbr);
 	if (flag->zero == 1)
 		flag->minus = -1;
@@ -35,7 +35,7 @@ char	*int_conditions(t_flags *flag, char *nbr)
 	{
 		nbr = fill_nbr_sign(nbr, flag->space);
 		if (nbr[0] == '+' || nbr[0] == ' ' || (nbr[0] == '-' && check == 1)
-			|| flag->precision > ft_strlen(nbr))
+			|| flag->precision > (int)ft_strlen(nbr))
 			flag->precision++;
 	}
 	return (nbr);
